Adding utils file for combined model functionality
